Time: 4 ms (46.7%), Space: 7.1 MB (18.45%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     bool judgeSquareSum(int c) {
         for(int a=0;a<=c;a++){
-            int t=c-(a*a);
+            int t=c-((long long)a*a);
             if(t<0)
                 break;
             int b=sqrt(t);
-            if((a*a+b*b)==c)
+            if(((long long)a*a+(long long)b*b)==c)
                 return true;
         }
         return false;
